+++ src/Service/LeantimeApiService.php
@@ -3,6 +3,7 @@
 namespace App\Service;
 
 use App\Exception\ApiServiceException;
+use App\Exception\EconomicsException;
 use App\Interface\DataProviderServiceInterface;
 use App\Model\Invoices\IssueData;
 use App\Model\Invoices\PagedResult;
@@ -29,11 +30,11 @@ use App\Model\SprintReport\SprintReportVersion;
 use App\Model\SprintReport\SprintReportVersions;
 use App\Model\SprintReport\SprintStateEnum;
 use Doctrine\Common\Collections\ArrayCollection;
+use Symfony\Component\Uid\Ulid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 class LeantimeApiService implements DataProviderServiceInterface
 {
-    private const PROJECT_TRACKER_IDENTIFIER = 'LEANTIME';
     private const API_PATH_JSONRPC = '/api/jsonrpc/';
     private const NO_SPRINT = 'NoSprint';
     private const PAST = 'PAST';
@@ -56,15 +57,10 @@ class LeantimeApiService implements DataProviderServiceInterface
         ];
     }
 
-    public function getProjectTrackerIdentifier(): string
-    {
-        return self::PROJECT_TRACKER_IDENTIFIER;
-    }
-
     /**
      * Get all projects, including archived.
      *
-     * @throws ApiServiceException
+     * @throws ApiServiceException|EconomicsException
      */
     public function getAllProjects(): mixed
     {
@@ -81,6 +77,7 @@ class LeantimeApiService implements DataProviderServiceInterface
      * @return mixed
      *
      * @throws ApiServiceException
+     * @throws EconomicsException
      */
     public function getProjectWorklogs($projectId, string $from = '2000-01-01', string $to = '3000-01-01'): mixed
     {
@@ -92,7 +89,7 @@ class LeantimeApiService implements DataProviderServiceInterface
      *
      * @return SprintReportProjects array of SprintReportProjects
      *
-     * @throws ApiServiceException
+     * @throws ApiServiceException|EconomicsException
      */
     public function getSprintReportProjects(): SprintReportProjects
     {
@@ -113,12 +110,11 @@ class LeantimeApiService implements DataProviderServiceInterface
     /**
      * Get projectV2.
      *
-     * @param $key
-     *   A project key or id
+     * @param string $projectId A project key or id
      *
      * @return SprintReportProject SprintReportProject
      *
-     * @throws ApiServiceException
+     * @throws ApiServiceException|EconomicsException
      */
     public function getSprintReportProject(string $projectId): SprintReportProject
     {
@@ -133,17 +129,27 @@ class LeantimeApiService implements DataProviderServiceInterface
 
     /**
      * @throws ApiServiceException
+     * @throws EconomicsException
      */
     private function getProjectIssuesPaged($projectId, $startAt, $maxResults = 50): array
     {
-        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['currentProject' => $projectId]);
+        // TODO: Implement pagination.
+        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => ['currentProject' => $projectId]]);
     }
 
+    /**
+     * @throws ApiServiceException
+     * @throws EconomicsException
+     */
     public function getIssuesDataForProjectPaged(string $projectId, $startAt = 0, $maxResults = 50): PagedResult
     {
         $result = [];
 
         $issues = $this->getProjectIssuesPaged($projectId, $startAt, $maxResults);
+
+        // Filter out all worklogs that do not belong to the project.
+        // TODO: Remove filter when issues are filtered correctly by projectId in the API.
+        $issues = array_filter($issues, fn ($issue) => $issue->projectId == $projectId);
 
         foreach ($issues as $issue) {
             $issueData = new IssueData();
@@ -151,19 +157,20 @@ class LeantimeApiService implements DataProviderServiceInterface
             $issueData->name = $issue->headline;
             $issueData->status = $issue->status;
             $issueData->projectTrackerId = $issue->id;
-            $issueData->projectTrackerKey = '';
+            // Leantime does not have a key for each issue.
+            $issueData->projectTrackerKey = $issue->id;
             $issueData->accountId = '';
             $issueData->accountKey = '';
-            $issueData->epicKey = $issue->tags ? $issue->tags : '';
-            $issueData->epicName = $issue->tags ? $issue->tags : '';
-            if ((bool) $issue->milestoneid && (bool) $issue->milestoneHeadline) {
+            $issueData->epicKey = $issue->tags;
+            $issueData->epicName = $issue->tags;
+            if (isset($issue->milestoneid) && isset($issue->milestoneHeadline)) {
                 $issueData->versions?->add(new VersionData($issue->milestoneid, $issue->milestoneHeadline));
             }
             $issueData->projectId = $issue->projectId;
             $result[] = $issueData;
         }
 
-        return new PagedResult($result, $startAt, $maxResults, count($issues));
+        return new PagedResult($result, $startAt, count($issues), count($issues));
     }
 
     public function getProjectDataCollection(): ProjectDataCollection
@@ -175,8 +182,9 @@ class LeantimeApiService implements DataProviderServiceInterface
             $projectData = new ProjectData();
             $projectData->name = $project->name;
             $projectData->projectTrackerId = $project->id;
-            $projectData->projectTrackerKey = '';
-            $projectData->projectTrackerProjectUrl = '';
+            // Leantime does not have a key for each project.
+            $projectData->projectTrackerKey = $project->id;
+            $projectData->projectTrackerProjectUrl = $this->leantimeUrl.'#/tickets/showTicket/'.$project->id;
 
             $projectVersions = $this->getSprintReportVersions($project->id);
             foreach ($projectVersions as $projectVersion) {
@@ -246,12 +254,25 @@ class LeantimeApiService implements DataProviderServiceInterface
     {
         $worklogDataCollection = new WorklogDataCollection();
         $worklogs = $this->getProjectWorklogs($projectId);
+
+        $workersData = $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.users.getAll');
+
+        $workers = array_reduce($workersData, function ($carry, $item) {
+            $carry[$item->id] = $item->username;
+
+            return $carry;
+        }, []);
+
+        // Filter out all worklogs that do not belong to the project.
+        // TODO: Remove filter when worklogs are filtered correctly by projectId in the API.
+        $worklogs = array_filter($worklogs, fn ($worklog) => $worklog->projectId == $projectId);
+
         foreach ($worklogs as $worklog) {
             $worklogData = new WorklogData();
-            if ((bool) $worklog->ticketId) {
+            if (isset($worklog->ticketId)) {
                 $worklogData->projectTrackerId = $worklog->id;
                 $worklogData->comment = $worklog->description ?? '';
-                $worklogData->worker = $worklog->userId;
+                $worklogData->worker = $workers[$worklog->userId] ?? $worklog->userId;
                 $worklogData->timeSpentSeconds = (int) ($worklog->hours * 60 * 60);
                 $worklogData->started = new \DateTime($worklog->workDate);
                 $worklogData->projectTrackerIsBilled = false;
@@ -259,11 +280,6 @@ class LeantimeApiService implements DataProviderServiceInterface
 
                 $worklogDataCollection->worklogData->add($worklogData);
             }
-
-            // TODO: Is this synchronization relevant?
-            // if (isset($worklog->attributes->_Billed_) && '_Billed_' == $worklog->attributes->_Billed_->key) {
-            //     $worklogData->projectTrackerIsBilled = 'true' == $worklog->attributes->_Billed_->value;
-            // }
         }
 
         return $worklogDataCollection;
@@ -735,7 +751,7 @@ class LeantimeApiService implements DataProviderServiceInterface
      */
     private function getIssuesForProjectMilestone($projectId, $milestoneId): array
     {
-        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['currentProject' => $projectId, 'milestone' => $milestoneId]);
+        return $this->request(self::API_PATH_JSONRPC, 'POST', 'leantime.rpc.tickets.getAll', ['searchCriteria' => ['currentProject' => $projectId, 'milestone' => $milestoneId]]);
     }
 
     /**
@@ -881,6 +897,7 @@ class LeantimeApiService implements DataProviderServiceInterface
      * Get from Leantime.
      *
      * @throws ApiServiceException
+     * @throws EconomicsException
      */
     private function request(string $path, string $type, string $method, array $params = []): mixed
     {
@@ -889,36 +906,25 @@ class LeantimeApiService implements DataProviderServiceInterface
                 ['json' => [
                     'jsonrpc' => '2.0',
                     'method' => $method,
-                    'id' => '1',
+                    'id' => (new Ulid())->jsonSerialize(),
                     'params' => $params,
                 ]]
             );
 
-            $body = $response->getContent(false);
+            $body = $response->getContent();
 
-            switch ($response->getStatusCode()) {
-                case 200:
-                    if ($body) {
-                        return json_decode($body, null, 512, JSON_THROW_ON_ERROR)->result;
-                    }
-                    break;
-                case 400:
-                case 401:
-                case 403:
-                case 409:
-                    if ($body) {
-                        $error = json_decode($body, null, 512, JSON_THROW_ON_ERROR);
-                        if (!empty($error->errorMessages)) {
-                            $msg = array_pop($error->errorMessages);
-                        } else {
-                            $msg = $error->errors->projectKey;
-                        }
-                        throw new ApiServiceException($msg);
-                    }
-                    break;
+            if ($body) {
+                $data = json_decode($body, null, 512, JSON_THROW_ON_ERROR);
+
+                if (isset($data->error)) {
+                    $message = $data->error->message;
+                    throw new ApiServiceException($message, $data->error->code);
+                }
+
+                return $data->result;
             }
         } catch (\Throwable $e) {
-            throw new ApiServiceException($e->getMessage(), (int) $e->getCode(), $e);
+            throw new ApiServiceException('Error from Leantime API: '.$e->getMessage(), (int) $e->getCode(), $e);
         }
 
         return null;
@@ -936,18 +942,5 @@ class LeantimeApiService implements DataProviderServiceInterface
         } else {
             return self::FUTURE;
         }
-    }
-
-    private function tagToId($tag)
-    {
-        // Use md5 hash function to generate a fixed-length hash
-        $hash = md5($tag);
-
-        // Extract three unique numbers from the hash
-        $num1 = hexdec(substr($hash, 0, 8)) % 1000;
-        $num2 = hexdec(substr($hash, 8, 8)) % 1000;
-        $num3 = hexdec(substr($hash, 16, 8)) % 1000;
-
-        return "$num1$num2$num3";
     }
 }